strava-elevation-addon/requirements.txt
numpy>=1.20.0
requests>=2.25.1
python-dotenv>=0.19.0
flask>=2.0.1
matplotlib>=3.4.0
pandas>=1.3.0
